refactor: make server cmd to start and migrate
