+++ cli/server.go
@@ -28,19 +28,34 @@ var (
 	Version string
 )
 
-func cmdServe() *cobra.Command {
-	return &cobra.Command{
-		Use:     "serve",
-		Short:   "Serve gRPC & HTTP service",
-		Long:    heredoc.Doc(`Serve gRPC & HTTP on a port defined in PORT env var.`),
-		Aliases: []string{"server", "start"},
+func serverCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "server <command>",
+		Aliases: []string{"s"},
+		Short:   "Run compass server",
+		Long:    "Server management commands.",
 		Example: heredoc.Doc(`
-			$ compass serve
+			$ compass server start
+			$ compass server start -c ./config.yaml
+			$ compass server migrate
+			$ compass server migrate -c ./config.yaml
 		`),
-		Args: cobra.NoArgs,
-		Annotations: map[string]string{
-			"group": "core",
-		},
+	}
+
+	cmd.AddCommand(
+		serverStartCommand(),
+		serverMigrateCommand(),
+	)
+
+	return cmd
+}
+
+func serverStartCommand() *cobra.Command {
+
+	c := &cobra.Command{
+		Use:     "start",
+		Short:   "Start server on default port 8080",
+		Example: "compass server start",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			cfg, err := loadConfig(cmd)
 			if err != nil {
@@ -49,6 +64,30 @@ func cmdServe() *cobra.Command {
 			return runServer(cfg)
 		},
 	}
+	return c
+}
+
+func serverMigrateCommand() *cobra.Command {
+	c := &cobra.Command{
+		Use:   "migrate",
+		Short: "Run storage migration",
+		Example: heredoc.Doc(`
+			$ compass migrate
+		`),
+		Args: cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			ctx, cancel := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
+			defer cancel()
+
+			cfg, err := loadConfig(cmd)
+			if err != nil {
+				return err
+			}
+
+			return runMigrations(ctx, cfg)
+		},
+	}
+	return c
 }
 
 func runServer(config Config) error {
@@ -183,4 +222,36 @@ func initNewRelicMonitor(config Config, logger log.Logger) (*newrelic.Applicatio
 	logger.Info("New Relic monitoring is enabled for", "config", config.NewRelic.AppName)
 
 	return app, nil
+}
+
+func runMigrations(ctx context.Context, config Config) error {
+	fmt.Println("Preparing migration...")
+
+	logger := initLogger(config.LogLevel)
+	logger.Info("compass is migrating", "version", Version)
+
+	logger.Info("Migrating Postgres...")
+	if err := migratePostgres(logger, config); err != nil {
+		return err
+	}
+	logger.Info("Migration Postgres done.")
+
+	return nil
+}
+
+func migratePostgres(logger log.Logger, config Config) (err error) {
+	logger.Info("Initiating Postgres client...")
+
+	pgClient, err := postgres.NewClient(config.DB)
+	if err != nil {
+		logger.Error("failed to prepare migration", "error", err)
+		return err
+	}
+
+	err = pgClient.Migrate(config.DB)
+	if err != nil {
+		return fmt.Errorf("problem with migration %w", err)
+	}
+
+	return nil
 }
